Simplifies invoice update process
Refactors the invoice update process by removing the custom logic for
handling client sessions. This simplifies the controller action
and relies on the model associations to manage client sessions.

Also changes the "Cancel" button on the edit invoice page to return to
the invoices index page rather than the invoice show page.