Adds support for third-party payees
Introduces the ability for invoices to be paid by someone other than the client.
This involves creating a new Payee model and modifying the Client and Invoice
models to support this new relationship. Also updates the UI to allow
selection of a payee for clients and invoices.

Addresses the scenario where a third party covers client session costs.