Refactors message client assignment
Builds on the standard Rails methods or assigning records to a record through a join table.  So we can write` message.clients = list_of_clients`.  The only difference is that if the list of clients is the string "all" we create the special record to show that this message should go to all clients.

Will simplify the controller.