Implements generic delete confirmation dialog
Implements a generic delete confirmation dialog using Stimulus.js.

This allows for a consistent user experience when deleting various entities (payees, clients, messages) throughout the application.  The controller now dispatches custom events with the entity type, which the delete modal controller listens for.

The delete button now uses data attributes to pass the ID and name of the entity, and determines entity type from its class name.

Also removes the need for the `confirm` popup when deleting.

Adds a test case to verify the functionality of the dialog and its cancellation.