Refactors invoice Paid
Replaces a dedicated invoices/mark_paid end point with just a standard PUT.
The Mark Paid button uses button_to send a PUT request to the controller's #update method to change the status to paid.
We have refactored the table so that the change to Paid simply replaces the row rather than the whole table.