Replaces duration with units for client sessions
This commit replaces the concept of "duration" in client sessions with "units".

Psychoanalysts tend to think in sessions rather than minutes.  They equate a session with an hour although the client is only seen for 50 mins and 10 mins are for write-up and comfort breaks.

Includes the migrations required to move from durations to units.