Changes how invoices can be billed directly to clients
For clients where a payee is setup, offers radio button to choose payee or client for invoice.

This change introduces a "self_paid" attribute to the Invoice model.
When set to true, the invoice is billed to the client directly,
bypassing the client's default payee. The UI has been updated
with radio buttons to toggle between billing to the payee and billing
to the client, when a payee is available.

Adds a new error to be raised when trying to set `self_paid` to true
when no payee is associated with the client.